requirements.txt
discord.py
jishaku
pylint
psutil